fix keep-alive thread never be closed
TransportImpl have a default serivce as nullService, when you set keepAlive to retain connection, a keep-alive thread will be create and periodic send msg to remote. but even after sshClient.close() has been called. this keepAlive will still exist. 
So, the solution is to set a ConnectionImpl object to TransportImpl.