Functional returns are easy to follow
This code base has a lot of instances where an
exception is thrown as part of the logic path.
That makes it very hard to follow and debug.

In Scala I use built in Types to encapsulate
the results of a method calls that may have
non-determinate results. I wrote something
like this version of PossibleResult for a
Java project recently, it used Optional
instead of null.

You may want to include this and use it,
or not.