This project isn't functional
This kind of thing might really make it easier to
follow the logic path, but as there is already a
huge base of exception throwing code, retrofitting
something like this would be a lot of work. Still
a little in the right spot might make things a lot
easier for a log time to come.